Move ZipWriter from common into web, because that's the only place it's used
