+++ onionshare/common.py
@@ -28,7 +28,6 @@ import sys
 import tempfile
 import threading
 import time
-import zipfile
 
 debug = False
 
@@ -224,54 +223,6 @@ def dir_size(start_path):
             if not os.path.islink(fp):
                 total_size += os.path.getsize(fp)
     return total_size
-
-
-class ZipWriter(object):
-    """
-    ZipWriter accepts files and directories and compresses them into a zip file
-    with. If a zip_filename is not passed in, it will use the default onionshare
-    filename.
-    """
-    def __init__(self, zip_filename=None, processed_size_callback=None):
-        if zip_filename:
-            self.zip_filename = zip_filename
-        else:
-            self.zip_filename = '{0:s}/onionshare_{1:s}.zip'.format(tempfile.mkdtemp(), random_string(4, 6))
-
-        self.z = zipfile.ZipFile(self.zip_filename, 'w', allowZip64=True)
-        self.processed_size_callback = processed_size_callback
-        if self.processed_size_callback is None:
-            self.processed_size_callback = lambda _: None
-        self._size = 0
-        self.processed_size_callback(self._size)
-
-    def add_file(self, filename):
-        """
-        Add a file to the zip archive.
-        """
-        self.z.write(filename, os.path.basename(filename), zipfile.ZIP_DEFLATED)
-        self._size += os.path.getsize(filename)
-        self.processed_size_callback(self._size)
-
-    def add_dir(self, filename):
-        """
-        Add a directory, and all of its children, to the zip archive.
-        """
-        dir_to_strip = os.path.dirname(filename.rstrip('/'))+'/'
-        for dirpath, dirnames, filenames in os.walk(filename):
-            for f in filenames:
-                full_filename = os.path.join(dirpath, f)
-                if not os.path.islink(full_filename):
-                    arc_filename = full_filename[len(dir_to_strip):]
-                    self.z.write(full_filename, arc_filename, zipfile.ZIP_DEFLATED)
-                    self._size += os.path.getsize(full_filename)
-                    self.processed_size_callback(self._size)
-
-    def close(self):
-        """
-        Close the zip archive.
-        """
-        self.z.close()
 
 
 class close_after_seconds(threading.Thread):
