+++ onionshare/web.py
@@ -26,6 +26,7 @@ import queue
 import socket
 import sys
 import tempfile
+import zipfile
 from distutils.version import LooseVersion as Version
 from urllib.request import urlopen
 
@@ -324,7 +325,7 @@ class Web(object):
         self.file_info['dirs'] = sorted(self.file_info['dirs'], key=lambda k: k['basename'])
 
         # zip up the files and folders
-        z = common.ZipWriter(processed_size_callback=processed_size_callback)
+        z = ZipWriter(processed_size_callback=processed_size_callback)
         for info in self.file_info['files']:
             z.add_file(info['filename'])
         for info in self.file_info['dirs']:
@@ -415,3 +416,51 @@ class Web(object):
                 urlopen('http://127.0.0.1:{0:d}/{1:s}/shutdown'.format(port, shutdown_slug)).read()
             except:
                 pass
+
+
+class ZipWriter(object):
+    """
+    ZipWriter accepts files and directories and compresses them into a zip file
+    with. If a zip_filename is not passed in, it will use the default onionshare
+    filename.
+    """
+    def __init__(self, zip_filename=None, processed_size_callback=None):
+        if zip_filename:
+            self.zip_filename = zip_filename
+        else:
+            self.zip_filename = '{0:s}/onionshare_{1:s}.zip'.format(tempfile.mkdtemp(), common.random_string(4, 6))
+
+        self.z = zipfile.ZipFile(self.zip_filename, 'w', allowZip64=True)
+        self.processed_size_callback = processed_size_callback
+        if self.processed_size_callback is None:
+            self.processed_size_callback = lambda _: None
+        self._size = 0
+        self.processed_size_callback(self._size)
+
+    def add_file(self, filename):
+        """
+        Add a file to the zip archive.
+        """
+        self.z.write(filename, os.path.basename(filename), zipfile.ZIP_DEFLATED)
+        self._size += os.path.getsize(filename)
+        self.processed_size_callback(self._size)
+
+    def add_dir(self, filename):
+        """
+        Add a directory, and all of its children, to the zip archive.
+        """
+        dir_to_strip = os.path.dirname(filename.rstrip('/'))+'/'
+        for dirpath, dirnames, filenames in os.walk(filename):
+            for f in filenames:
+                full_filename = os.path.join(dirpath, f)
+                if not os.path.islink(full_filename):
+                    arc_filename = full_filename[len(dir_to_strip):]
+                    self.z.write(full_filename, arc_filename, zipfile.ZIP_DEFLATED)
+                    self._size += os.path.getsize(full_filename)
+                    self.processed_size_callback(self._size)
+
+    def close(self):
+        """
+        Close the zip archive.
+        """
+        self.z.close()
